Not perfect, but moved things around and added tags
